app/requirements.txt
requests==2.9.1
Django==1.11.5
django_tastypie==0.14.0
